adding test cases for ccl and kv blocking

+++ tests/dynamo/unit_tests/test_continuous_batching.py
@@ -15,10 +15,11 @@ CPU-only. No QAIC hardware required.
 from __future__ import annotations
 
 import pytest
+import torch
 
 from QEfficient.transformers.models.modeling_auto import QEFFAutoModelForCausalLM
 
-from ._helpers import (
+from .._helpers import (
     DYNAMO_CAUSAL_LM_MODEL_IDS,
     assert_has_subfunctions,
     assert_retained_state_outputs,
@@ -37,7 +38,7 @@ def test_dynamo_cb_export(model_type, model_id, tmp_export_dir):
     """continuous_batching=True + dynamo=True and use_onnx_subfunctions=True."""
 
     try:
-        model_hf = load_hf_model(model_id)
+        model_hf = load_hf_model(model_id, torch_dtype=torch.float32)
     except Exception as exc:
         skip_on_model_fetch_error(exc, model_id)
